add tests
vet-all

rfctr + remove unused

fix regressions
