+++ integtests/chatbot-api/test_agent_conversation_flow.py
@@ -18,7 +18,12 @@ def available_agent(client: AppSyncClient):
     return agents[0]
 
 
-def _wait_for_session(client: AppSyncClient, session_id: str, expected_messages: int = 2, timeout: int = 30):
+def _wait_for_session(
+    client: AppSyncClient,
+    session_id: str,
+    expected_messages: int = 2,
+    timeout: int = 30,
+):
     """Wait for async message processing"""
     start_time = time.time()
     while time.time() - start_time < timeout:
@@ -29,13 +34,17 @@ def _wait_for_session(client: AppSyncClient, session_id: str, expected_messages:
         except Exception:
             pass
         time.sleep(1)
-    raise TimeoutError(f"Session {session_id} did not reach {expected_messages} messages within {timeout}s")
+    raise TimeoutError(
+        f"Session {session_id} did not reach {expected_messages} messages within {timeout}s"
+    )
 
 
-def test_agent_conversation_single_turn(client: AppSyncClient, available_agent, agent_session_id):
+def test_agent_conversation_single_turn(
+    client: AppSyncClient, available_agent, agent_session_id
+):
     """Test basic agent query → response flow"""
     agent_id = available_agent["agentRuntimeId"]
-    
+
     request = {
         "action": "run",
         "modelInterface": "agent",
@@ -53,28 +62,30 @@ def test_agent_conversation_single_turn(client: AppSyncClient, available_agent, 
                 "streaming": False,
                 "maxTokens": 2000,
                 "temperature": 0.7,
-                "topP": 0.9
-            }
-        }
+                "topP": 0.9,
+            },
+        },
     }
-    
+
     client.send_query(json.dumps(request))
     session = _wait_for_session(client, agent_session_id, expected_messages=2)
-    
+
     assert len(session["history"]) == 2
     assert session["history"][0]["type"] == "human"
     assert session["history"][1]["type"] == "ai"
-    
+
     # Verify AI response contains agent metadata
     ai_metadata = json.loads(session["history"][1]["metadata"])
     assert ai_metadata["provider"] == "bedrock-agents"
     assert ai_metadata["sessionId"] == agent_session_id
 
 
-def test_agent_conversation_multi_turn(client: AppSyncClient, available_agent, agent_session_id):
+def test_agent_conversation_multi_turn(
+    client: AppSyncClient, available_agent, agent_session_id
+):
     """Test conversation context preservation"""
     agent_id = available_agent["agentRuntimeId"]
-    
+
     # First message
     request1 = {
         "action": "run",
@@ -89,15 +100,20 @@ def test_agent_conversation_multi_turn(client: AppSyncClient, available_agent, a
             "videos": [],
             "modelName": "anthropic.claude-3-haiku-20240307-v1:0",
             "provider": "bedrock",
-            "modelKwargs": {"streaming": False, "maxTokens": 2000, "temperature": 0.7, "topP": 0.9}
-        }
+            "modelKwargs": {
+                "streaming": False,
+                "maxTokens": 2000,
+                "temperature": 0.7,
+                "topP": 0.9,
+            },
+        },
     }
-    
+
     client.send_query(json.dumps(request1))
     _wait_for_session(client, agent_session_id, expected_messages=2)
-    
+
     time.sleep(2)  # Brief delay between requests
-    
+
     # Follow-up message
     request2 = {
         "action": "run",
@@ -112,20 +128,24 @@ def test_agent_conversation_multi_turn(client: AppSyncClient, available_agent, a
             "videos": [],
             "modelName": "anthropic.claude-3-haiku-20240307-v1:0",
             "provider": "bedrock",
-            "modelKwargs": {"streaming": False, "maxTokens": 2000, "temperature": 0.7, "topP": 0.9}
-        }
+            "modelKwargs": {
+                "streaming": False,
+                "maxTokens": 2000,
+                "temperature": 0.7,
+                "topP": 0.9,
+            },
+        },
     }
-    
+
     client.send_query(json.dumps(request2))
     session = _wait_for_session(client, agent_session_id, expected_messages=4)
-    
+
     assert len(session["history"]) == 4
     # Verify message alternation: human → ai → human → ai
     assert session["history"][0]["type"] == "human"
     assert session["history"][1]["type"] == "ai"
     assert session["history"][2]["type"] == "human"
     assert session["history"][3]["type"] == "ai"
-
 
 
 def test_agent_conversation_error_handling(client: AppSyncClient, agent_session_id):
@@ -143,16 +163,21 @@ def test_agent_conversation_error_handling(client: AppSyncClient, agent_session_
             "videos": [],
             "modelName": "anthropic.claude-3-haiku-20240307-v1:0",
             "provider": "bedrock",
-            "modelKwargs": {"streaming": False, "maxTokens": 2000, "temperature": 0.7, "topP": 0.9}
-        }
+            "modelKwargs": {
+                "streaming": False,
+                "maxTokens": 2000,
+                "temperature": 0.7,
+                "topP": 0.9,
+            },
+        },
     }
-    
+
     # Should not raise exception
     try:
         client.send_query(json.dumps(request))
         # Wait briefly for error processing
         time.sleep(5)
-        
+
         # Check if session exists and has error handling
         session = client.get_session(agent_session_id)
         if session:
@@ -163,16 +188,18 @@ def test_agent_conversation_error_handling(client: AppSyncClient, agent_session_
         assert "invalid" in str(e).lower() or "not found" in str(e).lower()
 
 
-def test_agent_conversation_session_persistence(client: AppSyncClient, available_agent, agent_session_id):
+def test_agent_conversation_session_persistence(
+    client: AppSyncClient, available_agent, agent_session_id
+):
     """Test conversation persistence over time"""
     agent_id = available_agent["agentRuntimeId"]
-    
+
     messages = [
         "What is 2 + 2?",
-        "What is 5 * 3?", 
-        "Can you summarize our math conversation?"
+        "What is 5 * 3?",
+        "Can you summarize our math conversation?",
     ]
-    
+
     for i, message in enumerate(messages):
         request = {
             "action": "run",
@@ -186,22 +213,27 @@ def test_agent_conversation_session_persistence(client: AppSyncClient, available
                 "images": [],
                 "videos": [],
                 "modelName": "anthropic.claude-3-haiku-20240307-v1:0",
-            "provider": "bedrock",
-            "modelKwargs": {"streaming": False, "maxTokens": 2000, "temperature": 0.7, "topP": 0.9}
-            }
+                "provider": "bedrock",
+                "modelKwargs": {
+                    "streaming": False,
+                    "maxTokens": 2000,
+                    "temperature": 0.7,
+                    "topP": 0.9,
+                },
+            },
         }
-        
+
         client.send_query(json.dumps(request))
         expected_messages = (i + 1) * 2  # Each turn creates 2 messages
         _wait_for_session(client, agent_session_id, expected_messages=expected_messages)
-        
+
         if i < len(messages) - 1:  # Don't delay after last message
             time.sleep(2)
-    
+
     # Final verification
     session = client.get_session(agent_session_id)
     assert len(session["history"]) == 6  # 3 turns = 6 messages
-    
+
     # Verify message order and content
     for i in range(0, 6, 2):
         assert session["history"][i]["type"] == "human"
@@ -212,7 +244,7 @@ def test_agent_conversation_cleanup(client: AppSyncClient, available_agent):
     """Test session deletion"""
     cleanup_session_id = str(uuid.uuid4())
     agent_id = available_agent["agentRuntimeId"]
-    
+
     # Create conversation
     request = {
         "action": "run",
@@ -227,16 +259,21 @@ def test_agent_conversation_cleanup(client: AppSyncClient, available_agent):
             "videos": [],
             "modelName": "anthropic.claude-3-haiku-20240307-v1:0",
             "provider": "bedrock",
-            "modelKwargs": {"streaming": False, "maxTokens": 2000, "temperature": 0.7, "topP": 0.9}
-        }
+            "modelKwargs": {
+                "streaming": False,
+                "maxTokens": 2000,
+                "temperature": 0.7,
+                "topP": 0.9,
+            },
+        },
     }
-    
+
     client.send_query(json.dumps(request))
     _wait_for_session(client, cleanup_session_id, expected_messages=2)
-    
+
     # Delete session
     client.delete_session(cleanup_session_id)
-    
+
     # Verify session no longer exists
     time.sleep(2)
     try:
