--- lib/model-interfaces/bedrock-agents/__init__.py
@@ -1,1 +1,0 @@
-
