--- lib/model-interfaces/bedrock-agents/functions/__init__.py
@@ -1,1 +1,0 @@
-
