--- lib/model-interfaces/bedrock-agents/functions/request-handler/__init__.py
@@ -1,1 +1,0 @@
-
