+++ lib/shared/layers/python-sdk/python/genai_core/agents.py
@@ -1,5 +1,4 @@
 from typing import Any
-import json
 from datetime import datetime
 
 import genai_core.clients
@@ -17,19 +16,17 @@ def list_agents() -> list[dict[str, Any]]:
         list[dict[str, Any]]: List of agent information dictionaries
     """
     try:
-        client = genai_core.clients.get_bedrock_client(
-            service_name="bedrock-agentcore-control"
-        )
+        client = genai_core.clients.get_agentcore_control_client()
         response = client.list_agent_runtimes()
 
         agents = response.get("agentRuntimes", [])
-        
+
         # Convert datetime objects to strings for JSON serialization
         for agent in agents:
             for key, value in agent.items():
                 if isinstance(value, datetime):
                     agent[key] = value.isoformat()
-        
+
         return agents
     except ClientError as e:
         error_code = e.response.get("Error", {}).get("Code", "Unknown")
