+++ tests/shared/layers/python-sdk/genai_core/test_agents.py
@@ -1,10 +1,10 @@
-import pytest
 from unittest.mock import Mock, patch
-from botocore.exceptions import ClientError, BotoCoreError
+from botocore.exceptions import ClientError
 
 from genai_core.agents import list_agents
 
-@patch('genai_core.clients.get_bedrock_client')
+
+@patch("genai_core.clients.get_bedrock_client")
 def test_list_agents_success(mock_client):
     mock_bedrock = Mock()
     mock_bedrock.list_agent_runtimes.return_value = {
@@ -17,7 +17,8 @@ def test_list_agents_success(mock_client):
     assert result[0]["agentRuntimeId"] == "test-agent"
     mock_client.assert_called_with(service_name="bedrock-agentcore-control")
 
-@patch('genai_core.clients.get_bedrock_client')
+
+@patch("genai_core.clients.get_bedrock_client")
 def test_list_agents_client_error(mock_client):
     mock_bedrock = Mock()
     mock_bedrock.list_agent_runtimes.side_effect = ClientError(
